requirements.txt
numpy
matplotlib
tabulate
Pillow
imageio
torch
torchvision
geopandas
pygeos
shapely
